+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.1]
 
+### Features
+
+- Add share button
+
 ### Fixes
 
 - Fixed generate changelog for releases
+- Fixed mainfest errors
+- Fixed perforamance issues is reciter-list
 
 ## [0.2.0]